Rename DefaultMetricConsumers -> DefaultMetricsConsumer
